requirements.txt
django==1.9.9
django-guardian==1.4.5
django-import-export==0.4.5
mysqlclient==1.3.3